Allow registering different operand actors for different operands
