+++ mars/scheduler/operands/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 1999-2018 Alibaba Group Holding Ltd.
+# Copyright 1999-2019 Alibaba Group Holding Ltd.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,11 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .chunkmeta import ChunkMetaActor
-from .graph import GraphActor, GraphMetaActor
-from .operands import OperandActor, OperandState
-from .assigner import AssignerActor
-from .resource import ResourceActor
-from .session import SessionActor, SessionManagerActor
-from .kvstore import KVStoreActor
-from .utils import GraphState
+from .base import BaseOperandActor
+from .core import register_operand_class, get_operand_actor_class, \
+    OperandState, OperandPosition
+from .common import OperandActor
