+++ mars/scheduler/operands/tests/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 1999-2018 Alibaba Group Holding Ltd.
+# Copyright 1999-2019 Alibaba Group Holding Ltd.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,12 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from .chunkmeta import ChunkMetaActor
-from .graph import GraphActor, GraphMetaActor
-from .operands import OperandActor, OperandState
-from .assigner import AssignerActor
-from .resource import ResourceActor
-from .session import SessionActor, SessionManagerActor
-from .kvstore import KVStoreActor
-from .utils import GraphState
